# docs/swagger.yaml
basePath: /
definitions:
  entity.Operation:
    properties:
      date:
        type: string
      id:
        type: integer
      segment_name:
        type: string
      type:
        type: string
      user_id:
        type: integer
    type: object
...
  v1.addSegmentsByIdInput:
    properties:
      segment_ids:
        items:
          type: integer
        type: array
      ttl:
        type: string
      user_id:
        type: integer
    required:
    - segment_ids
    - user_id
    type: object
  v1.addSegmentsByNameInput:
    properties:
      segment_names:
        items:
          type: string
        type: array
      ttl:
        type: string
      user_id:
        type: integer
    required:
    - segment_names
    - user_id
    type: object
...
  v1.deleteSegmentsByIdInput:
    properties:
      segment_ids:
        items:
          type: integer
        type: array
      user_id:
        type: integer
    required:
    - segment_ids
    - user_id
    type: object
  v1.deleteSegmentsByNameInput:
    properties:
      segment_names:
        items:
          type: string
        type: array
      user_id:
        type: integer
    required:
    - segment_names
    - user_id
    type: object
  v1.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  v1.getOperationsHistoryInput:
    properties:
      month:
        type: integer
      page_size:
        type: integer
      user_ids:
        items:
          type: integer
        type: array
      year:
        type: integer
    required:
    - month
    - year
    type: object
  v1.getOperationsHistoryResponse:
    properties:
      operations:
        items:
          $ref: '#/definitions/entity.Operation'
        type: array
    type: object
  v1.getReportInput:
    properties:
      month:
        type: integer
      user_ids:
        items:
          type: integer
        type: array
      year:
        type: integer
    required:
    - month
    - year
    type: object
  v1.getReportLinkResponse:
    properties:
      link:
        type: string
    type: object
...
  v1.operationsResponse:
    properties:
      operation_ids:
        items:
          type: integer
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: msonkin33@gmail.com
    name: Maksim Sonkin
  description: This is a service for segmenting users with the ability to automatically
    add and remove users from segments.
  title: Dynamic User Segmentation Service
  version: "1.0"
paths:
  /api/v1/operations/add_segments/:
    post:
      consumes:
      - application/json
      description: addition a user to segments by id
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.addSegmentsByIdInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.operationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - Bearer: []
      summary: Add a user to segments by id
      tags:
      - operations
  /api/v1/operations/add_segments_by_names/:
    post:
      consumes:
      - application/json
      description: addition a user to segments by name
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.addSegmentsByNameInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.operationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - Bearer: []
      summary: Add a user to segments by name
      tags:
      - operations
  /api/v1/operations/delete_segments/:
    delete:
      consumes:
      - application/json
      description: delete user-segments relation by ids
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.deleteSegmentsByIdInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.operationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - Bearer: []
      summary: Delete User From Segments by ids
      tags:
      - operations
  /api/v1/operations/delete_segments_by_names/:
    delete:
      consumes:
      - application/json
      description: delete user-segments relation by names
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.deleteSegmentsByNameInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.operationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - Bearer: []
      summary: Delete User From Segments By Names
      tags:
      - operations
  /api/v1/operations/history/:
    get:
      consumes:
      - application/json
      description: getting operations history
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.getOperationsHistoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getOperationsHistoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - Bearer: []
      summary: Get Operations History
      tags:
      - operations
  /api/v1/reports/file/:
    get:
      consumes:
      - application/json
      description: getting history in csv format
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.getReportInput'
      produces:
      - application/json
      responses:
        "200":
          description: CSV-File
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - Bearer: []
      summary: Get CSV-File History
      tags:
      - reports
  /api/v1/reports/link/:
    get:
      consumes:
      - application/json
      description: getting history by link to csv-file
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.getReportInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getReportLinkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - Bearer: []
      summary: Get CSV-File Link
      tags:
      - reports
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"